Rename meshes file
